+++ test/models/PaymentSpec.scala
@@ -2,21 +2,23 @@ package models
 
 import org.scalacheck.Arbitrary
 import org.specs2.ScalaCheck
+import org.specs2.concurrent.ExecutionEnv
 import org.specs2.mutable.Specification
 import stellar.sdk.model.NativeAmount
-import stellar.sdk.model.op.PaymentOperation
 
-class PaymentSpec extends Specification with ScalaCheck {
+import scala.concurrent.Await
+import scala.concurrent.duration._
+
+class PaymentSpec(implicit ee: ExecutionEnv) extends Specification with ScalaCheck {
 
   implicit val arb: Arbitrary[Payment] = Arbitrary(Generators.genPayment)
 
   "a payment" should {
     "be convertible to a payment operation" >> prop { p: Payment =>
-      p.asOperation must beLike[PaymentOperation] { case op =>
-        op.amount mustEqual NativeAmount(p.units)
-        op.destinationAccount mustEqual p.destination
-        op.sourceAccount must beSome(p.source)
-      }
+      val op = Await.result(p.asOperation(), 30 seconds)
+      op.amount mustEqual NativeAmount(p.units)
+      op.destinationAccount mustEqual Await.result(p.destination.pk(), 5 seconds)
+      op.sourceAccount must beSome(Await.result(p.source.pk(), 5 seconds))
     }
   }